+++ app/src/main/java/com/localmusic/player/ui/home/HomeScreen.kt
@@ -420,7 +420,8 @@ fun HomeScreen(
                                             },
                                             onPrevious = onPrevious,
                                             onPlayPause = onPlayPause,
-                                            onNext = onNext
+                                            onNext = onNext,
+                                            onProgressChange = onProgressChange
                                     )
                                 }
                             }